+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add `YAML/Lint/DuplicateKey`.
 - Add opt-in `YAML/Style/KeyOrdering`.
 - Add opt-in `YAML/Style/KeyGrouping`.
+- Add opt-in local JSON Schema validation.
 
 —
 Stan Carver II